src/ingestion/metis/service_migration/requirements.txt
mysql-connector-python==8.0.24
pysftp==0.2.9
pandas==1.3.3